+++ resharper/resharper-unity/src/CSharp/Psi/CodeStyle/MemberReordering/EventFunction.cs
@@ -1,0 +1,29 @@
+using JetBrains.ProjectModel;
+using JetBrains.ReSharper.Psi.CSharp.Impl.CodeStyle.MemberReordering;
+using JetBrains.ReSharper.Psi.CSharp.Tree;
+using JetBrains.ReSharper.Psi.Tree;
+
+namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Psi.CodeStyle.MemberReordering
+{
+    public class EventFunction : INodeConstraint
+    {
+        public bool Matches(ITreeNode node)
+        {
+            if (node is IMethodDeclaration methodDeclaration)
+            {
+                var unityApi = node.GetSolution().GetComponent<UnityApi>();
+                return unityApi.IsEventFunction(methodDeclaration.DeclaredElement);
+            }
+
+            return false;
+        }
+
+        public int? Compare(INodeConstraint other)
+        {
+            if (other == null || other == Unconstrained.Instance)
+                return -1;
+
+            return null;
+        }
+    }
+}